scripts/requirements.txt
python-docx==1.1.2
requests==2.32.3
pandas==2.2.3